% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help statrt conversions}
\usage{
hi(who)
}
\arguments{
\item{who}{character(1) The name of the person to greet at}
}
\value{
a line of the text to be used for starting the conversion
}
\description{
This function is for greeting in different ways
}
\examples{
   hi("Song Liu")
 
}

